// src/app/data/types/ProjectData.ts
////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////

import { IconLabelKey } from "@/app/data/enums/IconLabelKey";

////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////

export type ProjectData = {
  title: string;
  imageSrc: string;
  imageAlt: string;
  videoSrc: string;
  iframeSrc: string;
  summary: string;
  description: string;
  repositoryLinkText: string;
  repositoryLinkHref: string;
  iconLabelKeyArray: IconLabelKey[];
}

////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////